requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
datasets==4.1.1
dill==0.4.0
fasttext-wheel==0.9.2
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
gensim==4.3.3
huggingface-hub==0.34.4
idna==3.10
intel-openmp==2021.4.0
Jinja2==3.1.6
joblib==1.5.2
kagglehub==0.3.13
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
mkl==2021.4.0
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
networkx==3.4.2
numpy==1.26.4
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
pyarrow==21.0.0
pybind11==3.0.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.13.1
six==1.17.0
smart_open==7.3.0.post1
sympy==1.14.0
tbb==2021.11.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.67.1
transformers==4.56.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wrapt==1.17.3
xformers==0.0.32.post2
xxhash==3.5.0
yarl==1.20.1
